Removes the use of `vugu/js` on the GOOS=js GOARCH=wasm target
This commit removes the use of `vugu/js` for the GOOS=js and GOARCH=wasm
target. It replaces `vugu/js` with a direct reference to `syscall/js` for
this target platform.

For all other platforms the code still uses the `vugu/js` package via
the `vugu/js/impl-nonjs.go` file.

Potentially the file `vugu/js/impl-js.go` could now be removed.

Summary of the changes:

Any code that uses the `js` package has been split.
On the GOOS=js and GOARCH=wasm platform this code is placed in
a file named `component_js_wasm.go` and imports `syscall/js`.
No build tag is added to this file. The `_js_wasm.go` filename tells the
Go compiler to only build this file on a GOOS=js and GOARCH=wasm
platform.

For all other platforms any code that referneces `js.Xxxx` is pulled out
to a file named `component_notjs_notwasm.go`. As the file extension has
no meaning to the Go tool chain this file also has a build tag of

//go:build !js || !wasm

added to it.

You can see this change in the `domrender` package, the `events-dom.go
file and the `vgnode.go` file.

Any file with a `*.go` extension is now assumed to be a common file used
on all platforms.

parser.Parser.Go has been changed to output two files.

The first file is an unchanged `component_gen.go` file but with an added
build tag (as above). This is NOT correct - see note below. This file
shoudl be named `component_gen_notjs_notwasm.go`.

The second file is a `component_gen_js_wasm.go` file that has no build
tag and will only be built on a GOOS=js GOARCH=wasm platform.

Note: The `component_gen.go` file should be called
`component_gen_notjs_notwasm.go` in line with the code naming above.
However doing this breaks the `gen/parser` test set. The reason for this
is unclear but appears to be related to the structure of the tests
themselves. As a result the files named `component_gen.go` do have a
build tag of:

//go:build !js || !wasm

and as such should not be considered common code. These `*_gen.go` files are
actually any platform except GOOS=js GOARCH=wasm.

The parser tests will need to be refactored at some future point to
support the correct naming convention of
`component_gen_notjs_notwasm.go` and the parser will need to write out a
file of this name.