Fix: Ensure we run a `go mod tidy` after we do a dependency update
A dependency update calls `go get -u -t` for each module in vugu. But
while this adds newer versions of dependencies into the `go.mod` and
`go.sum` it does not remove old dependencies i.e. older versions of
modules.

The only way to remove older module versions from the `go.mod` and
`go.sum` is to run `go mod tidy` on the same module.

This is behaviour is made clear in the `go mod tidy` documentation.

This commit ensures that a `go mod tidy` is always run if the proceeding
`go get -u -t` did not error.