Fix: remove all the wasm_exec.js files.
This commit removes all of the `wasm_exec.js` files from the repo, and
updates the `magefile` so that we do not check that the `wasm_exec.js`
file for differences - we can't because it is no longer in the repo.

The change does not remove the copy of `wasm_exec.js` from the
`tinygo-dev` package or from the `legacy-wasm-test-suite`. It also does
not add a `wasm_exec.js` into the project wide `.gitignore` file. This
can only be done if the `legacy-wasm-test-suite` and the `tinygo-dev`
packages are both removed.

This partially closes Issue #344m see:

https://github.com/vugu/vugu/issues/344

The issue is only partially closed because we have not yet implemented
`tinygo` support in the `magefiles`. We will either need new targets
that are specific to using `tinygo` or a flag or envvar(?) to indicate
that `tinygo` should be used.

Until this is resolved we don't know how to change the `magefile` to
discover the value of the $GOROOT env var for `tinygo`. This points to a
different location relative to the standard Go compiler. We need to know
the updated value to ensure we copy the `tinygo` version of
the `wasm_exec.js` file and not the standard Go version.