Copy the wasm_exec.js file from the correct location pre and post go1.24
Go 1.24 changed the location of the wasm_exec.js file from:

$GOROOT/misc/wasm
to
$GOROOT/lib/wasm

See the relase notes for go1.24.

As we currently still build using go1.23 (the current oldstable
release) the magefiles need to be updated to account for this change.

This change can be removed once go.1.25 arrives in Aug 2025. At that
point both the stable and oldstable releases will used the newer
$GOROOT/lib/wasm directory.