Remove the compiler tests from the devutil package
Removes the tests of the standard Go and tinygo compilers from the
devutil package.

These tests test that:

a) the compiler generates valid wasm, but that is the job of the
compiler teams
b) the "build system" defined in the files compiler.go and
tinygo-compiler.go work as expected - but calling the compiler is the
responsibility of the built system, and not the go of the project
itself.

Removing these tests removes a test failure caused by the tinygo compiler
not (yet) fully suporrting the Go 1.24.0 changes.

The remains of the devutil package have been left as is as they form
part of the current public API. See Issue #281.

Removing these files then requires that the simplehttp/simple-handler.go
file is updated to take into account the new location of the
wasm_exec.js file from Go v1.24.0 onwards. This change can be removed
once Go 1.25.0 is released.

The change ensures that for go v1.24.0 or greater the new location is
used, for versions prior to this the old location of wasm_exec.js is
used.