Fix the staticrender ValueOf test
The staticrender ValueOf tests will only ever be executed on a GOOS=!js
|| GOARCH=!wasm platform i.e, on the native platform only.

Because the test uses the limited, but cross platform vugu/js nonjs
implementation of the package we expect the test to panic.

The test has now been updated to expect the panic string.

If this test were to fail in the future it would indicate that the wrong
vugu/js implementation has been called.