Remove the vg-for.noshadow option
Go 1.22 (beginning in Go 1.21 under the GOEXPERIMENT=loopvar flag)
the behaviour of the for loop changed. Prior to Go 1.22 loop variables
where scoped per loop. From Go 1.22 onwards they are scoped perloop
iteration.

This removes the need for the `vg-for.noshadow` option. The behaviour
of the `vg-for` and `vg-for.noshadow` are now idential.

The generated code will still insert shadowed variables in the body
of the for loop, but this be a no-op assignment to self which the Go compiler will optimise away.