Remove the shadowed loop variables from the generated code
The parser has been simplified to remove the shadowed loop variables
from the generated code.

This greatly simplifies the `(p *ParserGo)emitForExpr` method. The method
now only emits a Go `for ....` statement which can be range, iteration
or C style.

The cleanup included removing the local `iterval` and `iterkey` variables
as well the `vgiterkeyx`` concept. None of these concepts are now
necessary.

The test code has been updated to reflect these changes.
The test code however is brittle, as it relies on the line ending with
a `\n` in the expected output.