Bug Fix: Update the vg-for example after the noshadow option was removed
Updating the vg-for example after the noshadow vg-for option has been
removed reveals a serious bug.

Previously the `root.vugu` file contained this:

<!-- This example uses the explicit form of the vg-for -->
<!-- This time the methods c.DaysInWeekIthSorted uses the new iter, sllice and maps package internally -->
  <li vg-for='index, value := range c.DaysInWeekIterator()'>
    <div vg-content='fmt.Sprintf("The day of the week at index %d is %s", index, c.DayOfWeek(index))'></div>
  </li>

This vg-for statement should have failed to compile. But it did not.
It should have failed with an unused valriable "value"
in the `root_gen.go` file.

This compilation failure is correct - the variable "value"
declared in the loop declaration is not used in the body of the loop.

Following the removal of the `noshadow` option this test code then
(correctly) failed with the unused variable error.

The test code has now been corrected to remove any reference to the
unused variable, "value".