Ensure auto generated shortcut loop variables are alwasy referenced
In the case of a `vg-for` shortcut like this:

<li vg-for='c.Items()'> ... </li>

It is impossible to tell if the two auto generated loop variables
"key" and "value" are ever used as part of a `vg-content` attribute of
the tag itself (in this case the <li> tag) or by some other tag
enclosed by this <li> tag.

In the case were either of the auto generated loop variables are not used the
Go compiler then correctly fails to build the auto generated code because
one of variables "key" or "value" is unused.

But, without changing the parser so that it looks ahead to see which (if
any) of the auto generated variables are used in an enclosed scope it is
impossible to tell what the minimal code is that should be generated.

As a compromise solution, in the case where the `vg-for` shortcut style
is used the parser now always generates a pair of unused variables like
so:

_ = key
_ = value

as the first two lines inside the loop body.

This is sufficient to prevent the compiler generating an "decalred But
unused variable" error.