Fix the gen packages tests and the wasm-test-suite tests
The gen.Parser.Run() method no longer generates a file named "component_gen.go".
Instead it generates a pair of files, "component_gen_js_wasm.go" and
"component_gen_notjs_notwasm.go".

This commit updates the gen package tests so that the package tests now
pass, using the new filenames. Previously the tests treated the
"component_gen.go" file as the native i.e. !js || !wasm file, but
this is a fudge. Although the file name implies it is a cross platform
file this is in fact false, as the embedded build tag of "//go:build !js
|| !wasm" proved.

This therefore brings the tests which run on the native platform into
line with the wasm-test-suite of tests which run on a GOOS=js
GOARCH=wasm platform i.e. in a browser.

The changes to the tests are fairly minimal.

In missing_fixer_test.go in test TestMissingFixer the test now writes
out a file called "root_gen_notjs_notwasm.go" with an build tag of
"//go:build !js || !wasm". This replaces the previous "root_gen.go" file
that was created. The missingFixer struct has also been updated to
look for a file named "root_gen_notjs_notwasm.go" that is file generated
from the "roo.vugu" file.

In parser-go-pkg.go in function Run the missingFmap (missing file names
map?) has been updated to look for a file named
"root_gen_notjs_notwasm.go". The function previously looked for a file
called "root_gen,go", this is because the ParserGoPkg.Run() method scans
the package directory and looks for *.vugu files. In this case it will
find a "root.vugu" file. Previously, it then constructed a filename of
the form "<basename>"+"_gen"+".go". But this clearly does not take
into account the platform specific naming. This is now set explicitly to
the native platform file name.

In parser-pg-pkg_test.go in test TestSimpleParseGoPkgRun the test now
looks for a file named "root_gen_notjs_notwasm.go" and not "root_gen.go"
as was previously.

In the same file in test TestRun the test case "Simple" now generates an
output file called "root_gen_notjs_notwasm.go", that includes a the
correct build tag.

In parser-go.go in function writeNonJsPlatformFile the function now
correctly writes out a fill with a "_gen_notjs_notwasm.go" suffix and
extension. Previously it used the filename as passed in the parameter
which would have have a "_gen.go" suffix and extension. This now
parallels the behaviour of the writeJsPlatformFile function.

All of the gen package tests are designed to run on the native platform,
and not in a browser. As  result they only look for and use the
"*_gen_notjs_notwasm.go" files.

All other changes in the gen paclage are to add debug logging/tracing.
These will be removed prior to the final merge of this PR.

In the wasm-test-suite the test file "main_test.go" for test-021-slots has
changed. Although test test runs on the native platform it has been updated
to ensure both the "root_gen_notjs_notwasm.go" and "root_gen_js_wasm.go"
files are both present on the local disk.

The remaining changes are all file renames under the "wasm-test-suite"
and "vgform" directories.

The renaming is always of the form of:

"some_component_gem.go" => "some_component_gen_notjs_notwasm.go"

This final rename change is required. Without it the "Build(...)"
function will be defined in both the old "component_gen.go" file and
the new "component_gen_notjs_notwasm.go" file on the native
platform resulting in a compilation failure.