Initial vugu command that only supports the "version" sub command
An initial version of the proposed `vugu` command, from Issue #340
https://github.com/vugu/vugu/issues/340

The only supported sub command is "version" which prints the version
information. For example:

vugu version v0.4.0-139-g3fe0108-dirty

The magefiles have been updated so that the `build` target now
automatically inserts the correct version information into the
`vugu` binary, based on the output of a `git describe --dirty` command.

If the `vugu` tool is built outside of the magefile e.g. with "go build
github.com/vugu/vugu/cmd/vugu" or otherwise outside of the repo then the version information only contains
the full hash of the last commit and an indication if the repo was
dirty.

The different formats allow the build method to be inferred.

The `vugu` tool is built with the "github.com/urfave/cli/v3" package,
which simplifies the development of the sub commands, flag parsing and help information.