Add the fetch-tags option to the checkout action
By default the checkout action does not checkout the repos tags, and only
checks out the last commit.

This causes a failure because the magefile calls `git describe --dirty`
to determine the correct version string, which is relative to the last
tag.

If there are no tags, the `git describe --dirty` fails which in turn
causes the build to fail.

The checkout action not explictly checks out the tags.