Add support for "-v" and "--version" flags that are provided by default
The cli package supports both a "-v" and a "--version" flag by default.
In addition the `vugu` command will print its help information, which
includes its version number, if no arguments are passed.

The first two cases require setting the cli.VersionPrinter function
to the version.Version function.

The last case requires that the version information is read at
startup. If we fail to read the built in build information at this
point we report an error and exit. Once the version has been read we
need to chop off the "vugu version" prefix. This ensures that when the
version information is displayed as part of the vugu commands help it is
displayed correctly. When called via the "-v", "--version" or "version"
sub command the prefix is always displayed.